Separate out burden estimate writing to specialist class.
